Updated drand48 driver_c for unsigned long long iterations
Since the RAND_MAX of drand48 is 2^32 - 1 (at most) an int did not have
enough bits to store the cycle length of the drand48 algorthim.  By
updating it to unsigned long long I can generate at least 1 full
iteration of the alogrthim.

+++ drand48/driver_c/generate.c
@@ -1,5 +1,6 @@
 #include <stdio.h>
 #include <stdlib.h>
+#include <stdint.h>
 
 int main( int argc, char *argv[] ) {
     if ( argc != 3 ) {
@@ -7,13 +8,12 @@ int main( int argc, char *argv[] ) {
         return 1;
     }
 
-    int seed           = atoi( argv[1] );
-    int num_iterations = atoi( argv[2] );
-
+    int seed = atoi( argv[1] );
+    unsigned long long num_iterations = atoll( argv[2] );
 
     srand48(seed);
 
-    int iteration;
+    unsigned long long iteration;
     for ( iteration = 0; iteration < num_iterations; iteration++) {
         printf("%f\n", drand48());
     }
